+++ tests/logs/__init__.py
@@ -1,3 +1,5 @@
+"""This package contains unit tests for logging."""
+
 #  Copyright (c) 2022 EPAM Systems
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -10,11 +12,3 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License
-
-"""This package contains core reportportal-client modules."""
-
-from reportportal_client.logs import log_manager
-
-__all__ = [
-    'log_manager'
-]